+++ contracts/handlers/curve/ILiquidityGauge.sol
@@ -9,5 +9,6 @@ interface ILiquidityGauge {
     function set_approve_deposit(address addr, bool can_deposit) external;
     function deposit(uint256 _value, address addr) external;
     function withdraw(uint256 _value) external;
+    function transfer(address _to, uint256 _value) external returns (bool);
     function claimable_tokens(address addr) external view returns (uint256);
 }